9_ML_System_Monitoring/monitoring_proj/app/requirements.txt
fastapi
uvicorn
tensorflow
pillow
prometheus_client
python-multipart